This change adds sorting to the comments.
Reddit can fail to organize a website by old for the full duration of the comments. Luckily, each comment ID is already created, from 0 (being the oldest) to z (being the youngest). Thanks to this, the “comment_sort = ‘old’” part of the code may no longer be needed. I’ll try a new collection in the next five minutes.